+++ core-ng/src/test/java/core/framework/impl/validate/BeanValidatorNotBlankTest.java
@@ -1,7 +1,8 @@
 package core.framework.impl.validate;
 
-import core.framework.api.validate.NotEmpty;
+import core.framework.api.validate.NotBlank;
 import core.framework.api.validate.NotNull;
+import core.framework.util.ClasspathResources;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -14,12 +15,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * @author neo
  */
-class BeanValidatorNotEmptyTest {
+class BeanValidatorNotBlankTest {
     private BeanValidator validator;
+    private BeanValidatorBuilder builder;
 
     @BeforeEach
     void createObjectValidator() {
-        validator = new BeanValidatorBuilder(Bean.class, Field::getName).build().orElseThrow();
+        builder = new BeanValidatorBuilder(Bean.class, Field::getName);
+        validator = builder.build().orElseThrow();
     }
 
     @Test
@@ -33,6 +36,12 @@ class BeanValidatorNotEmptyTest {
         assertTrue(errors.hasError());
         assertEquals(1, errors.errors.size());
         assertThat(errors.errors.get("stringField1")).contains("stringField1");
+    }
+
+    @Test
+    void sourceCode() {
+        String sourceCode = builder.builder.sourceCode();
+        assertEquals(ClasspathResources.text("validator-test/validator-not-blank.java"), sourceCode);
     }
 
     @Test
@@ -50,10 +59,10 @@ class BeanValidatorNotEmptyTest {
 
     static class Bean {
         @NotNull
-        @NotEmpty(message = "stringField1 must not be empty")
+        @NotBlank(message = "stringField1 must not be blank")
         public String stringField1;
 
-        @NotEmpty(message = "stringField2 must not be empty")
+        @NotBlank(message = "stringField2 must not be blank")
         public String stringField2;
     }
 }