+++ core-ng/src/test/java/core/framework/impl/web/service/TestWebService.java
@@ -5,7 +5,6 @@ import core.framework.api.json.Property;
 import core.framework.api.validate.Length;
 import core.framework.api.validate.Max;
 import core.framework.api.validate.Min;
-import core.framework.api.validate.NotEmpty;
 import core.framework.api.validate.NotNull;
 import core.framework.api.validate.Pattern;
 import core.framework.api.validate.Size;
@@ -62,7 +61,6 @@ public interface TestWebService {
 
     class TestRequest {
         @NotNull
-        @NotEmpty
         @Pattern("\\d+.*")
         @Length(max = 10)
         @Property(name = "string_field")