+++ core-ng-test/src/test/java/core/framework/test/assertion/ValidatorAssertTest.java
@@ -1,6 +1,7 @@
 package core.framework.test.assertion;
 
 import core.framework.api.validate.Length;
+import core.framework.api.validate.NotBlank;
 import core.framework.api.validate.NotNull;
 import org.junit.jupiter.api.Test;
 
@@ -13,13 +14,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 class ValidatorAssertTest {
     @Test
     void failWithMessage() {
-        Bean bean = new Bean();
+        var bean = new Bean();
         bean.field2 = "";
 
         assertThatThrownBy(() -> assertBean(bean).isValid())
                 .isInstanceOf(AssertionError.class)
                 .hasMessageContaining("to be valid, but found some violations:")
-                .hasMessageContaining("{field1=field must not be null, field2=field must not be empty}");
+                .hasMessageContaining("{field1=field must not be null, field2=field must not be blank}");
     }
 
     @Test
@@ -41,6 +42,7 @@ class ValidatorAssertTest {
         @Length(max = 5, message = "field1 must not be longer than 5")
         public String field1;
 
+        @NotBlank
         public String field2;
     }
 }